zctas bounding boxes from tiger
"The Census Bureau has trademarks on a number of our TIGER-related product names*. The purpose of trademarking these terms is so that customers buying products with those names can be confident in identifying the origin (the Census Bureau in this case) of the products (data) being sold. That shouldn't dissuade anyone from creating products using the TIGER/Line data or even citing our original products as the source of all or part of their product. In fact, many of our products are resold or redistributed by others. Vendors take the basic product, add value to it (snazzier interface, more data, etc.) and resell it. Sometimes they just resell exact copies of the data. That's all perfectly legal and we are happy to see the data redistributed and get wider usage."