Don't delete common words found in parsed phrases.
The first thing we do when we retry parsing is remove any common words
(e.g. "city", "country", etc.) from the list of tokens.  This has the
unfortunate side effect of making it impossible to match features that
have those words in their name.

For example, the token list ['kansas', 'city'] will be reduced to just
['kansas'].  That will match Kansas (the state) instead of Kansas City
(the city), which is probably not what was originally intended.

We now only remove words from the token list that don't appear in the
set of parsed phrases.