limit max radius to 5km for now
with the new circle code, increasing the radius increases the amount of work you need to do a lot. For now, cap it at 5km because at high neighborhood density, that's about what you can answer in 300ms.

this is a hack that I don't feel great about putting in the public codebase, btu we have so many internal callers that don't sanitize their radius.

Auditors: nsanch, rahul