start of refactoring importers to be more pluggable for TGN
the idea being that there can be multiple parsers that implement a basic interface that yields GeocodeRecords.

this kills the "geometry" column parsing because that code is terrible - we should fix this in our hacked input file generation side. it is really awful.
it makes PolygonLoader rely on the  IndexRunner object which is not great and hopefully will be fixed

the "loadIntoMongo" is I think the wrong interface to expose -- my idea for now is make a helper class that knows how to take a GeocodeRecord, do all the name transforms on it, and save it to the database and have GeonamesParser use that. Then see if I can easily modify it so that the parser just generates a stream of GeocodeRecords and some parent class does the actual databse insertions

Auditors: rahul