Fix hotfix adds to parentIds from clobbering old parentIds.
The issue here is that the GeocodeServingFeature for SoHo seems to have Nil parentIds on the GeocodeFeature field.
The parentIds seem to be on the scoringFeatures field, though.
This leads me to suspect it's an issue with how we are building the index, since they should always be the same.