Move MapFileConcurrentReader into twofishes pkg
MapFileConcurrentReader previously was in
org.apache.hadoop.io because it tried to reuse
MapFile's comparator setting.  However, this
setting was never set, and with the CDH5 upgrade,
the access of ComparatorOption was more restricted

Removing the custom comparator, which was never used,
allowed the class to be moved into twofishes where it belongs

If we ever need to actually use a custom comparator, we don't
need to use MapFile.Reader's.  We can just subclass Option ourselves.

Auditors: dan